adapter/obd/j1939.cpp: safer treatment of `size_` property.
Message size in TP.CM_RTS request is 16 bit wide - it may take up to 4 hex digits.

+++ src/adapter/obd/j1939connmgr.h
@@ -28,7 +28,7 @@ public:
     bool isValidAck(const CanMsgBuffer* msg) const;
     bool rts(const CanMsgBuffer* msg);
     bool data(const CanMsgBuffer* msg);
-    uint32_t size() const { return size_; }
+    uint16_t size() const { return size_; }
 private:
     bool sendAck();
